+++ lib/commands/fortnite/locker/image.dart
@@ -1,5 +1,3 @@
-import "dart:convert";
-
 import "package:nyxx/nyxx.dart";
 import "package:nyxx_commands/nyxx_commands.dart";
 import "package:nyxx_interactions/nyxx_interactions.dart";
@@ -28,6 +26,7 @@ final ChatCommand lockerImageCommand = ChatCommand(
     (
       IContext ctx, [
       @Description("User to view locker for") IUser? user,
+      @Description("PNG quality") bool png = false,
     ]) async {
       if (client.cachedCosmetics.isEmpty) {
         throw Exception(
@@ -96,13 +95,12 @@ final ChatCommand lockerImageCommand = ChatCommand(
       List<int> img;
       for (var i = 0; i < chunks.length; i++) {
         int startTime = DateTime.now().millisecondsSinceEpoch;
-        img = base64Decode(
-          await drawLocker(
-            cosmetics: chunks[i],
-            epicname: dbUser.fnClient.displayName,
-            username:
-                "${(user ?? ctx.user).tag} [${chunks[i].length}/${cosmetics.length}]",
-          ),
+        img = await drawLocker(
+          cosmetics: chunks[i],
+          epicname: dbUser.fnClient.displayName,
+          username:
+              "${(user ?? ctx.user).tag} [${chunks[i].length}/${cosmetics.length}]",
+          png: png,
         );
 
         await ctx.respond(